Added an interface for non-primitive datatypes.
Essentially the 'double' datatype is replaced by the
type 'su2double'. If nothing is specified during the
configuration, this is just a typedef to double. Otherwise
it uses a user-defined type. As a demonstration a complex
type is implemented (for example for complex step
derivatives). If --enable-complex is given as an option to
the configure script, a variant of SU2_CFD (called
SU2_CFD_COMPLEX) is build that utilizes this complex datatype.