Make NTS and FD numerics blending more consistent
Before, Dissipation=0 indicated an upwind scheme for FD but a central
scheme for NTS, and vice versa for Dissipation=1. The difference made
both the code and the output confusing. This commit changes the
`Roe_Dissipation` variable so that 0 = central and 1 = upwind for both
schemes.

Some code cleaning and precondition checking were also added.