Add a max cell length function to Geometry
DES and DDES rely on the max grid spacing for a cell. Previously, this
was computed for every iteration, which was inefficient.  By moving this
calculation to the geometry classes, the max grid spacing will only be
calculated whenever the geometry changes.

The more complex DDES grid-length variants (SA_ZDES and SA_EDDES) rely
on flow variables, and consequently *must* be calculated in the flow
solver, once per iteration.