Change dynamic viscosity to kinematic viscosity
The original definition of `r_d` in DES used kinematic viscosity (not
dynamic viscosity).  Using dynamic viscosity is not dimensionally
consistent, and will lead to problems if density is not 1.  Since
`GetLaminarViscosity()` and `GetEddyViscosity()` return the dynamic
viscosity, it is necessary to divide by the density to get kinematic
viscosity.