Add a max cell length function to CGeometry
DES and DDES rely on the max grid spacing for a cell.  Previously, this
was computed for every iteration, which was inefficient.  There were
also duplication, with at least 4 different places where the max
cell length was calculated. By moving this calculation to the geometry
classes, the max grid spacing will only be calculated whenever the
geometry changes.

The more complex DDES grid-length variants (SA_ZDES and SA_EDDES) rely
on the flow variables, and consequently *must* be calculated in the flow
solver, once per iteration.