Adjust inlet profile filename for no restart cases
When using a custom inlet profile using the C++-only method (as opposed
to python), the inlet profiles require iteration numbers just like the
restart files.  This creates a problem when an inlet file is used but
the solution is started fresh (with no restart). Before, the solver
would error-out with a message stating that SU2 was requesting a
negative iteration.

This commit forces the iteration number used when loading inlet profiles
to be 0 when the simulation is *not* restarted.