Move viscous numerics to a new header file
Each time the numerics header file is modified, all the dependencies
must be recompiled.  Unfortunately, this is pretty much the whole
program.  By moving the viscous numerics to its own header file, only
the code that depends on the viscous numerics (which is very little)
must be recompiled with each change.