Change "Corrected" classes to a boolean option
Using different classes to provide different behavior is sometimes the
best way.  However, in this case, about 95% of the code is duplicated
between the "Corrected" class for the viscous numerics (e.g.
CAvgGradCorrected_Flow), and only 5% varies.  This could be fixed using
a template design pattern, but the extra complexity would be
unnecessary.  Instead, a simple boolean switch is sufficient.