Change floating-point math to make tests pass
This commit introduces two small changes which do not change the
analytical equations, but do change the calculated values due to
floating point roudoff errors.  This commit is **not** meant to be part
of the final production code.  Rather, it serves to illustrate that
changing these two sections of code changes the floating-point roundoff
errors in an incompressible CHT simulation run under non-ideal
configuration settings.