Adjust CHT test case to allow for more tolerance.
Since the CHT incompressible test case is not run long enough to reach a
reasonable convergence, the heat flux value is both very large and
transient at the final iteration of the test. It was noticed that the
final value during testing is sensitive to small changes in the
floating-point arithmetic (e.g. ordering of operations).  By increasing
the tolerance, the test is less likely to fail when bug-free changes to
the code are made.

For more information, see the discussion on PR #626 of su2code/SU2.