Fix max length MPI comm. for periodic geometries
Some of the max lengths on halo nodes are not properly set when using
periodic BCs.  A quick hack is to call Set_MPI_MaxLength twice, to
ensure that all the cells on both send and receive boundaries get the
max length.  A better code fix would require restructuing how the
periodic boundaries work.