Allow invalid sensor values in convective blending
If the "DUCROS" versions of convective central/upwind blending is
specified, the Ducros sensor values are needed.  But if a "DUCROS"
version (i.e. NTS_DUCROS) is *not* specified, then the sensor values are
never used.  The Roe low dissipation function should only check the
sensor values if they're actually going to be used.  Otherwise, it
should allow invalid (or unitialized) values to pass through.