Add missing term to strain magnitude for 2D
When summing over the diagonal components of the deviatoric
rate-of-strain tensor, only the first two diagonal components are summed
in 2D.  This creates problems, because the third component is *not*
zero.  The third component of the rate-of-strain tensor in a 2D is zero
(i.e. S_33 = 0).  But the deviatoric rate-of-strain tensor has a nonzero
third diagonal component.  So in 2D, the third diagonal component must be
explicitly added.

This should only make significant impacts on problems where the
divergence of velocity is non-negligible when compared to the diagonal
components of the rate-of-strain tensor.