Removed MARKER_INTERFACE option.
MARKER_INTERFACE was not used anymore. The flag INTERFACE_BOUNDARY is
now used to initialize boundaries set with MARKER_ZONE_INTERFACE. Their
specific kind of BC will be set during the Interface_Preprocessing() in
the driver. This removes the necessity of using MARKER_CHT_INTERFACE or
MARKER_FLUID_INTERFACE in addition to MARKER_ZONE_INTERFACE.