Add aero coefficients to output when required
In the revamp of output in v7.0.0, it looks like a subtle bug was
introduced.  If the config file is setup using default settings,
the aerodynamic coefficients are *not* output. This is true even if
the python function `aerodynamics` is called.  The end result
is that aerodynamics is called to compute the aero coefficients, but
the coefficients aren't stored anywhere.

This creates a fatal bug in trying to restart shape optimization.