Update streamwise periodic reg test values.
PR#1022 SIMD introduced some minor difference in my(!) cht reg tests.
8184779..4b9f2a8x contains #1022 & #1080 (only 5 lines).
The change is in solid only and only affects the 2D case. Not the 3D.
I dont know what specifically introduced the changes, but as they
are small I for now assume that it is just a little numeric change.