Fixed Coord_j for boundary viscous numerics
The visc_numerics' coordinates were often set to the coordinates
of the boundary vertex and its closest normal neighbour. Setting
them to the boundary vertex and the reflection of the closest
normal neighbour at the vertex is more systematic however.
Normally this should not change much, because the boundary numerics
do not apply gradient correction, so they only depend on the
distance between Coord_i, Coord_j.