correct array sizes in CInterpolator::Collect_VertexInfo
Before, Buffer_Send_Coord was
a) resized to nVertexDonor rows (with nDim entries each). Then
b) the number of local domain vertices is the number of rows
   written. Finally,
c) MaxLocalVertex_Donor rows were communicated.
The number in (b) is less then the numbers in (a) and (c), so we
do not write beyond scope and we communicate everything that has
been written. However, in the communication we might access non-
reserved memory, which is undefined behaviour AFAIK? Therefore,
we correct the number in (a) to match the number in (c).