Fix pseudo time update, conv Jacobians, diff Jacobians.
The transported variable is rho*Y but the Solution stores just Y.
Jacobians have to be taken with respect to the transported variable
rho*Y.
The linear system computes an update to rho*Y but we have transform that
to fit the solvers Solution Y (by dividing by rho).
The SST-solver is basically a perfect baseline because it transports
rho*k and rho*w.