Fix the solver restarts, especially for turb+scalar combined.
The scalar solver is handeled last of the solvers. It updates species
fractions which alter mean-flow fluid-properties. Therefore a Flow-Pre
followed by a Turb-Post has to be done, which makes the correct eddy
visc available.
At restarts Scalar-Pre has to be called as Post is empty and the
Turb-Post has to be called after Flow-Pre due to the same reasonsas
above.