Rename scalar files to scalarLegacy to avoid files with the same name.
PR 1330 introduces a general scalar framework that has files with the
same name (i.e. CScalarSolver) but that have slightely different scopes.
Therefore a renaming of certain current files with the *Legacy addition
is done to avoid merge conflicts. Merging #1330 which is already in
develop means that we keep up with develop which is wanted.
This means we will introduce the 'new' species/scalar solver into
feature_flamelet with the old one still present.