Changes to Testcases cfg's. Fix convergence for inc cases.
Mostly some cosmetic changes. Added all residual fields and Linear
Solver stuff. CFL for inc cases was too high to properly converge. This
will break the regression test so that needs to be changed in a future
commit.
An open questions that needs to be answered is which input values are
given to the case: Re, Ma, mu_lam, etc? And which ones are free.