Add Direct Solution reset for Heat Singlezone.
Before an empty implementation was called for this such that the
direct solution was not reset to the original one after the
CLEAR_INDICES recording. Only affected HEAT SingleZone which was quite
likely not used very often. The MultizoneDriver calls the solver routine
directly instead of going over the iteration->SetRecording.