DAMultizone-driver now calls iteration->SetRecording.
Instead of calling the solver->Recording directly. For Heat and Fea this
is no difference at all as the iteration does nothing else than calling
the solver routine. For Fluid there is an additional loop over iMesh for
the FLOW_SOL and additional SetRecordings for Turb, WeaklyHeat,
Radiation. Those are handeled by the multizone in the old implementation
as well.