Make ENUM_TURB_MODEL an enum class.
Went well for the most part. In some switches a default: break; had to
be added. in option_structure.inl the old ENUM_TURB_MODEL was stored in
some unsigned short intInfo container which is not seamlessly possible
as an enum class names its own new type. Imo the storage is not not
necessary as it is never used so I just removed it.