Fix multiple inlet-file reading. Especially the template write.
So problem was a classic mismatch between what iMarkers are loop over in
the function that sets sth and the other that writes sth.
At one point there is an alphabetical sort of marker tags which destroys
the order given by the config which implicitly can be used to stay
consistent.
The sort was removed and during template inlet writing the same iMarker
loop is used to be consistent with the value preparation.