Add rans/naca0012 testcases for explicit Euler
adapted from turb_NACA0012_sst_fixedvalues.cfg
I will fill in the residuals of the CI pipeline in the next commit.
Note that these tests only check that the behavior of the explicit
Euler scheme is constant, but not that it is correct (i.e., it is
the explicit Euler scheme) at the time of this commit.