Fill dtor for species-inlet in option_structure.inl
Note that the other options that also allocate memory on the heap via
new do not specifically call a delete in their dtor. But they also only
need 1d arrays and species needs 2d arrays. Maybe that is a difference.
That dtor in opt_struct.inl is called via the CConfig dtor at its very
start.