Added AvgSurface Species output in SetAnalyzeSurface function.
The values are not made acessible to the outside via config like the
other quantities. That is used for Obj.Funcs

A "bug" was noticed: For multiple analyze markers the values from each
marker are just added up for the total. This is legit for e.g. massflow
but not for e.g. Temperature where one gets an avg temp of 900 with 3
outlets that have an avg of 300...