Add Species Variance output.
A bit ugly AFTER the other Marker_analyze quantities. This is necessary
because we need the avg_Species for the variance computation.
The Variance vlaue is computed using the specific marker avg and per
marker and NOT for each species but accumulated. I.e. if one needs very
good mixing between 2 speies and the third is not important then this
additional complexity needs to introduced.