Add dp for streamwise massflow to flow-metadata to allow restarts with that setting.
Note though that the comment in the pressure update computation of the streamwise periodicity
still holds: Currently the pressure update for massflow is omitted in the first iteration,
as the function is called in the Preprocessing phase which leads to multiple calls during 1 iteration.
Ideally this whole method is called in another place or a setup has to be found which prevents multiple calls.