Remove DV_MARKER check for multiple zones.
CConfig checks whether the DV_MARKERs appear in the marker list. But for multizone
cases it can happen, that the marker to be deformed are just in the zonal configs.
The check would fail although everything is alright.
To check it right one would need to check this after all zonal cfgs were read and then
check it once with a loop over all cfgs to see whether DV_MARKER appears somewhere there.