Add Integrated Heatflux to Inc/Comp flow as well.
Therefore a SurfaceArea routine was added to CGeometry which is called once
in the GeoPreproc of the Driver. This central Surface area comp makes the
HF change straight forward and might be useful in some other cases.
The ordering of the SurfaceArea vector of CGeometry follows the Global cfg ordering.