Adapt a reg test to contain INTEGRATED_HEATFLUX.
Used a 3D cht case and applied it in the solid zone. Residual stay exactly the same.
This, to my experience is only the case because the surface is a plane rectangle.
For e.g. a round shape the small mismatch in exact circle circumference and the discrete
appriximation by the mesh is enough to alter the residuals (and final temperature)
slighthly. The Residual change in this case for the Temperature is due to the error that
was in the mesh (see Testcases #91)