Move SP pressure drop update to the Source_residual.
Before this update routine was called together with the NS Preprocessing.
This is troublesome as the Preprocessing is called multiple times during startup and
special precaution has to taken in order to avoid that the update happens multiple times (which,
by the way, happened anyway as the FlowPre is called before and after the MultiGridCycle).
Now in the Source Residual it is called directly before its usage which means no crazy logic is necessary
to prevent writing during multiple Pre steps and thus, makes sure that influence is taped during the adjoint as well.
The adjoint is still not working for a PressureDrop OF (the available one that is then computed
using the recovered Pressure which shouldn't be an issue because the Pre computes the latest one after
the iteration). Drag and AvgT look okayish but grad differences ~50% are to be expected.