Treat streamw.per. dp like a solution var in the adjoint solver.
As during the primal solution-finding process there is an iteration in the
pressure drop for prescribed massflow, the pressure drop should be treated like
the Solution as well. That means:
0. Storing of the restarted value to reset for the multiple recordings
1. Register Input and(!) output and therefore storage of the AD_Index
2. Set the AD-adjoint of the output with the Lagrange-Adjoint of the pressure drop (initVal 1e-16)
3. Extract the AD-adjoint of the input dp and set it as the Lagrange-Adjoint

I am not really set on how to treat multiple processes. Currently I just Allreduce
the local contributions at the AD-adjoint extraction. Needs testing.
But this version in single core is able to validate AvgTemp for single-zone massflow prescription
which wasn't possible before.