Make pressure-drop adjoint work with mpi. Do not allreduce Adjoint_dp.
Seemingly, the adjoint_dp can stay local to each core and does not need
to be accumulated across processes via Allreduce during adjoint extration.
This version works well with multiple processes and gives accurate gradientes against
FD.

I am not fully sure whether I understand why it works without communication.
1. The solution of course only really lives on 1 core as that is the point
of domain decomposition.
2. For fixed Variables like AoA, the allreduce is done, AoA of course equally lives on
all processes.
The streamwise periodic pressure drop also equally lives on all processes. But my guess, why
the extra comm is not needed is: The communication for dp is done in the primal solver. So
there the "link" between the dp's on all cores is made and is not required explicitely again. The
same argument would be used for the solution vars.
For the variables like AoA this is not the case. For the AD-tool, AoA really lives on each
process alone without all of them beeing linked, that is why the allreduce is necessary.