First push for correct Multizone pressure-drop adjoint. Not workin yet.
For true multizone cases (nZone>1) the cross-term evaluation writes those sensitivities
into the Solution container. Therefore the "true" adjoint Solution needs to be
stored away intermediately in order to reset that after the cross-term eval.
That intermediate container ins Solution_BGS_k. The same concept in the same
places is here applied for Adjoint_DP. The code does the right thing but the sens
are still wrong so there has to be another mistake.

Testing is done with a single zone but MULTIZONE=YES but it doesnt work there either.