Implement Extra adjoint equation in a more generic way.
Now the user has to provide the adress of each extra adjoint var in the primal
solver. The primal solver tracks those and the rest of the code should be generic.

Unfortunately there is no push_back() for the C2DContainer class .. using just std::vector
would solve this remedy.