Fix running averages for restarted simulations.
When restarted at e.g. Time-Iter 2 the first value would be divided by 3 (number of timesteps counting from zero).
So if your first value in that restarted sim is a 3, the running average would be 1.

This is fixed with this commit by setting WINDOW_START_ITER = RESTART_ITER in the Config Postprocessing,
if the option is not set. If the option is set, but incorrectly, an error is raised.