allow to use system-wide Meson and Ninja
See github issue [#1945] for more information on the rationale.

[#1945]: <https://github.com/su2code/SU2/issues/1945>

If a user wants to build SU2 from source and prefers to use
system-wide Meson and Ninja (assuming they are installed on the OS),
rather than SU2-specific copies, it is possible to avoid downloading
SU2-specific versions:

  $ SU2_MESON='no' ./meson.py
  $ export CXXFLAGS=<preferred_compiler_options>
  $ meson setup <preferred_config_options> build
  $ ninja -C build install

The default behavior is unchanged. If a user does not have system-wide
Meson and Ninja installed on the OS and/or prefers to use SU2-specific
copies, it is stll possible to download and use SU2-specific versions
of Meson and Ninja:

  $ export CXXFLAGS=<preferred_compiler_options>
  $ ./meson.py setup <preferred_config_options> build
  $ ./ninja -C build install